Generalize calculations
Previous calculations worked when the parent of the frozen object was scaled
up and down, but the calculations didn't account for any scaling of grandparents,
etc.  This uses the parent worldTransform scale values, which allows us to add
intermediate layers between the world root and our frozen objects.  The benefit of
this is increased control over the order in which objects are rendered on the canvas
(eg render order is no longer just object creation order)